services/orders/requirements.txt
Django
djangorestframework
djongo
psycopg2
requests